Work on a reader + trampolining.
I want to split Tokenizer into a Reader and a 'Parser'. That way I can
write a scheme interpreter in scheme.

I want proper tail calls. I have decided to implement it with
trampolining. I will use the fact that scheme procedures are first class
functions. I'm not sure it's feasible / possible to do the trampolining
in a mix of java and scheme so I think I will do it all in scheme by
adding a bounce type and some kind of while loop.